add sschema module, with extensibility
Instead of each project that wants to use schemas having to write their
own schema checking logic, we can put the core logic into a module and
allow users to leverage it. In addition, we provide a method of adding
your own jsonschema formatcheckers and handlers, as well as providing
some prebuilt ones in the sschema module.